A ROWS or GROUPS frame bound is typed like LIMIT but gives a bare
placeholder no clause name. The analyze_builtin_macros case casts the
fdiv divisor: with the overload rule of this merge a bare placeholder
between fdiv's double and float overloads is ambiguous.

+++ internal/endtoend/testdata/analyze_builtin_macros/duckdb/query.sql
@@ -23,7 +23,7 @@ FROM runs;
 SELECT id FROM runs
 WHERE list_contains(array_append(retry_reasons, $reason), 'timeout')
   AND split_part(reason, $sep, 1) = 'timeout'
-  AND fdiv(n, $divisor) > 1;
+  AND fdiv(n, $divisor::DOUBLE) > 1;
 
 -- name: AggregateMacros :one
 SELECT json_group_array(reason) AS reasons, geomean(id) AS geo, wavg(n, id) AS weighted FROM runs;
